requirements.txt
fastapi
uvicorn
google-generativeai
pytesseract
pdfplumber
pillow
numpy
requests
pydantic